feat: Update ConfirmSwapModal
